#2708 fixed invalid query bug and moved team type apis into respective folders
