+++ src/frontend/src/apis/design-reviews.api.ts
@@ -2,7 +2,7 @@
  * This file is part of NER's FinishLine and licensed under GNU AGPLv3.
  * See the LICENSE file in the repository root folder for details.
  */
-import { CreateTeamTypePayload, EditDesignReviewPayload } from '../hooks/design-reviews.hooks';
+import { EditDesignReviewPayload } from '../hooks/design-reviews.hooks';
 import axios from '../utils/axios';
 import { DesignReview } from 'shared';
 import { apiUrls } from '../utils/urls';
@@ -24,19 +24,6 @@ export const getAllDesignReviews = () => {
   return axios.get(apiUrls.designReviews(), {
     transformResponse: (data) => JSON.parse(data).map(designReviewTransformer)
   });
-};
-
-/**
- * Gets all the team types
- */
-export const getAllTeamTypes = () => {
-  return axios.get(apiUrls.allTeamTypes(), {
-    transformResponse: (data) => JSON.parse(data)
-  });
-};
-
-export const createTeamType = async (payload: CreateTeamTypePayload) => {
-  return axios.post(apiUrls.teamTypesCreate(), payload);
 };
 
 /**
