+++ src/frontend/src/hooks/teams.hooks.ts
@@ -4,7 +4,7 @@
  */
 
 import { useQuery, useQueryClient, useMutation } from 'react-query';
-import { Team, TeamType } from 'shared';
+import { Team } from 'shared';
 import {
   getAllTeams,
   getSingleTeam,
@@ -15,8 +15,6 @@ import {
   createTeam,
   setTeamLeads
 } from '../apis/teams.api';
-import { CreateTeamTypePayload } from './design-reviews.hooks';
-import { editTeamType } from '../apis/team-types.api';
 
 export interface CreateTeamPayload {
   teamName: string;
@@ -131,22 +129,6 @@ export const useSetTeamLeads = (teamId: string) => {
     {
       onSuccess: () => {
         queryClient.invalidateQueries(['teams']);
-      }
-    }
-  );
-};
-
-export const useEditTeamType = (teamTypeId: string) => {
-  const queryClient = useQueryClient();
-  return useMutation<TeamType, Error, CreateTeamTypePayload>(
-    ['team-type', 'edit'],
-    async (formData: CreateTeamTypePayload) => {
-      const { data } = await editTeamType(teamTypeId, formData);
-      return data;
-    },
-    {
-      onSuccess: () => {
-        queryClient.invalidateQueries(['team-type', 'teams']);
       }
     }
   );
