+++ src/frontend/src/pages/CalendarPage/CalendarPage.tsx
@@ -15,9 +15,9 @@ import { useAllDesignReviews } from '../../hooks/design-reviews.hooks';
 import ErrorPage from '../ErrorPage';
 import { useCurrentUser } from '../../hooks/users.hooks';
 import { datePipe } from '../../utils/pipes';
-import { useAllTeamTypes } from '../../hooks/design-reviews.hooks';
 import LoadingIndicator from '../../components/LoadingIndicator';
 import DRCSummaryModal from './DesignReviewSummaryModal';
+import { useAllTeamTypes } from '../../hooks/team-types.hooks';
 
 const CalendarPage = () => {
   const theme = useTheme();
