+++ src/frontend/src/pages/TeamsPage/SetTeamTypeModal.tsx
@@ -7,8 +7,7 @@ import LoadingIndicator from '../../components/LoadingIndicator';
 import ErrorPage from '../ErrorPage';
 import NERFormModal from '../../components/NERFormModal';
 import { FormControl, FormLabel, Select, MenuItem } from '@mui/material';
-import { useAllTeamTypes } from '../../hooks/design-reviews.hooks';
-import { useSetTeamType } from '../../hooks/team-types.hooks';
+import { useAllTeamTypes, useSetTeamType } from '../../hooks/team-types.hooks';
 
 interface SetTeamTypeInputs {
   teamId: string;
